requirements.txt
google-cloud-storage
pandas
scikit-learn==0.20.0
joblib
matplotlib
